refactor: String을 List<LottoNumber>로 변환하는 메서드를 LottoTicket으로 이동
