+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -62,6 +62,6 @@ class LottoTicketTest {
                 .map(Integer::parseInt)
                 .map(LottoNumber::new)
                 .collect(Collectors.toList());
-        return new LottoTicket(lottoNumbers);
+        return LottoTicket.fromSixNumbers(lottoNumbers);
     }
 }
